+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Change the datasource url to be the multi-tenant-proxy in front of the loki-gateway.
+- Add secret management for the proxy by duplicating loki-auth.
+
 ### Fixed
 
 - Fix incorrect alloy security context.